fix: Tiger Style compliance - unit suffixes and field naming
Additional Tiger Style naming compliance across 83 files:
- capacity->capacity_tokens for rate limiters
- capacity->capacity_permits for semaphores
- capacity->capacity_jobs for job pools
- max->max_bytes in error types (DecompressionTooLarge, TokenTooLarge)
- Fixed test assertions using old field names (size->size_bytes,
  enabled->is_enabled, deleted->is_deleted)
- Fixed property test to use calculate_available_capacity_f32
- Fixed DLQ test using deprecated should_move_to_dlq signature
- Added clippy allow attributes for complex return types and many args

Library builds successfully with all features.

+++ crates/aspen-auth/src/builder.rs
@@ -119,7 +119,7 @@ impl TokenBuilder {
         // Validate capability count
         if self.capabilities.len() > MAX_CAPABILITIES_PER_TOKEN as usize {
             return Err(AuthError::TooManyCapabilities {
-                count: self.capabilities.len(),
+                count: self.capabilities.len() as u32,
                 max: MAX_CAPABILITIES_PER_TOKEN,
             });
         }
